Split write worker controller into module
