+++ lib/writable-controller.js
@@ -1,7 +1,8 @@
-const { spawn } = require('child_process');
 const stream = require('stream');
-const { writeToFileDir } = require('./util/fs');
+
+const WriteWorkerController = require('./write-worker-controller');
 const { SQLBuffer, BUFFERS } = require('./sql-buffer');
+const { writeToFileDir } = require('./util/fs');
 
 
 const {
@@ -51,60 +52,25 @@ function buildSchema(commandContents) {
 }
 
 
-function sendToWorker(worker, obj, cb = () => {}) {
-  const str = typeof obj === 'string' ? obj : JSON.stringify(obj);
-  const configBuffer = Buffer.from(`${str}\n`);
-
-  worker.stdin.write(configBuffer, cb);
-}
-
-
 // Primary Class Definition
 
 class MultiWriteable extends stream.Writable {
-  constructor(config = {}) {
+  constructor(config = {
+    numWorkers: 1,
+    outputDir: null,
+    schemaFormat: null,
+  }) {
     super();
 
-    const {
-      numWorkers,
-      outputDir,
-      schemaFormat,
-    } = Object.assign({}, config);
+    Object.assign(this, config);
 
     this.buffer = new SQLBuffer();
+    this.writeWorkers = new WriteWorkerController(config);
 
     this.tables = {};
     this.currentTable = null;
 
-    this.outputDir = outputDir;
-    this.schemaFormat = schemaFormat;
-
-    this.numWorkers = parseInt(numWorkers, 10);
-    this.workers = [];
-    this.readyWorkers = [];
-
-    for (let i = 0; i < this.numWorkers; i += 1) {
-      const worker = spawn('node',
-        [`${__dirname}/worker`], {
-          detached: true,
-          stdio: ['pipe', 'inherit', 'inherit'],
-        });
-
-      worker
-        .on('error', (err) => {
-          throw err;
-        })
-        .on('exit', (exitCode) => {
-          if (exitCode !== 0) {
-            console.error(`Worker ${i}-PID ${worker.pid} exited with code: ${exitCode}`);
-          }
-        });
-
-      this.workers[i] = worker;
-      this.readyWorkers[i] = worker;
-    }
-
-    this.sendConfig();
+    this.writeWorkers.sendConfig();
   }
 
   _write(chunk, enc, next) {
@@ -195,6 +161,7 @@ class MultiWriteable extends stream.Writable {
 
   continueInsert() {
     const writes = [];
+    const { readyWorkers } = this.writeWorkers;
     let row;
     let worker;
     let queueItem = this.buffer.getNextCommandParenSet();
@@ -202,11 +169,11 @@ class MultiWriteable extends stream.Writable {
     while (
       queueItem
       && queueItem !== COMMAND_EXEC_BUFF
-      && this.readyWorkers.length
+      && readyWorkers.length
     ) {
-      worker = this.readyWorkers.pop();
+      worker = readyWorkers.pop();
       row = `[${queueItem}]`;
-      writes.push(this.writeToNextAvailableWorker(worker, row));
+      writes.push(this.writeWorkers.writeToNextAvailableWorker(worker, row));
 
       queueItem = this.buffer.getNextCommandParenSet();
     }
@@ -270,7 +237,7 @@ class MultiWriteable extends stream.Writable {
         };
 
         this.tables[tableName] = schema;
-        this.sendSchema(schema);
+        this.writeWorkers.sendSchema(schema);
 
         if (this.schemaFormat) {
           this.writeSchema(schema);
@@ -302,51 +269,6 @@ class MultiWriteable extends stream.Writable {
       filename,
       content: JSON.stringify(schemaColumns, null, 2),
     });
-  }
-
-
-  // //////////////////////////////////////////////////////
-  // Worker Methods
-
-  writeToNextAvailableWorker(worker, row) {
-    const numWorkersAvailable = this.readyWorkers.length;
-
-    return new Promise((resolve) => {
-      let callNext = resolve;
-
-      if (numWorkersAvailable) {
-        callNext = () => {};
-        resolve();
-      }
-
-      sendToWorker(worker, row, () => {
-        this.readyWorkers.push(worker);
-        callNext();
-      });
-    });
-  }
-
-  sendConfig() {
-    const workerConfig = {
-      outputDir: this.outputDir,
-      totalWorkers: this.workers.length,
-    };
-
-    this.workers
-      .forEach((worker, i) => {
-        const config = Object.assign({
-          workerPart: i,
-        }, workerConfig);
-
-        sendToWorker(worker, { config });
-      });
-  }
-
-  sendSchema(schema) {
-    this.workers
-      .forEach((worker) => {
-        sendToWorker(worker, { schema });
-      });
   }
 }
